move notification component to AppBanner dir
