+++ client/src/App.tsx
@@ -1,9 +1,5 @@
 import Container from "@mui/material/Container";
 import PathwayScreen from "./components/pathwaysTab/pathwayScreen";
-import SplashScreen from "./components/SplashScreen/SplashScreen";
-import AccountsScreen from "./components/AccountsTab/AccountsScreen";
-
-import CaregiverScreen from "./components/CareGivers/CaregiverScreen";
 
 function App() {
